[refactor] 모듈 이동 mono -> search-domain-es
