+++ domain/search-domain-es/src/main/java/woowa/team4/bff/search/domain/MenuSearch.java
@@ -1,5 +1,6 @@
 package woowa.team4.bff.search.domain;
 
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -7,12 +8,14 @@ import lombok.Getter;
 @Getter
 @Builder
 @AllArgsConstructor
-public class RestaurantSearch {
+public class MenuSearch {
+
     private String id;
-    private String restaurantName;
+    private String menuName;
+    private Long menuId;
     private Long restaurantId;
 
-    public void update(String restaurantName) {
-        this.restaurantName = restaurantName;
+    public void update(String menuName) {
+        this.menuName = menuName;
     }
 }
