+++ domain/search-domain-es/src/main/java/woowa/team4/bff/search/repository/MenuSearchRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 import woowa.team4.bff.search.document.MenuSearchDocument;
 
 public interface MenuSearchRepository extends ElasticsearchRepository<MenuSearchDocument, String> {
-    MenuSearchDocument findByMenuId(Long menuId);
 
+    MenuSearchDocument findByMenuId(Long menuId);
     List<MenuSearchDocument> findByMenuNameContaining(String menuName);
 }
