+++ domain/search-domain-es/src/main/java/woowa/team4/bff/search/repository/RestaurantSearchMapper.java
@@ -5,8 +5,10 @@ import org.mapstruct.factory.Mappers;
 import woowa.team4.bff.search.document.RestaurantSearchDocument;
 import woowa.team4.bff.search.domain.RestaurantSearch;
 
+
 @Mapper
 public interface RestaurantSearchMapper {
+
     RestaurantSearchMapper INSTANCE = Mappers.getMapper(RestaurantSearchMapper.class);
     RestaurantSearch toDomain(RestaurantSearchDocument document);
     RestaurantSearchDocument toDocument(RestaurantSearch domain);
