+++ domain/search-domain-es/src/main/java/woowa/team4/bff/search/repository/SearchIndexManageRepository.java
@@ -8,6 +8,7 @@ import woowa.team4.bff.search.domain.RestaurantSearch;
 @Repository
 @RequiredArgsConstructor
 public class SearchIndexManageRepository {
+
     private final RestaurantSearchRepository restaurantSearchRepository;
     private final MenuSearchRepository menuSearchRepository;
     private final RestaurantSearchMapper restaurantSearchMapper = RestaurantSearchMapper.INSTANCE;
