--- mono/src/main/java/woowa/team4/bff/search/command/UpdateRestaurantSearchCommand.java
@@ -1,5 +1,0 @@
-package woowa.team4.bff.search.command;
-
-public record UpdateRestaurantSearchCommand(Long restaurantId, String restaurantName) {
-}
-
