--- mono/src/main/java/woowa/team4/bff/search/domain/RestaurantSearchResult.java
@@ -1,12 +1,0 @@
-package woowa.team4.bff.search.domain;
-
-public class RestaurantSearchResult {
-    private String restaurantUuid;
-
-    private String restaurantName;
-    private Integer minimumOrderPrice;
-
-    private Double rating;
-    private Integer reviewCount;
-    private String menuName;
-}
